small changes in handler, chain, and create /process
