+++ package.json
@@ -26,6 +26,7 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@types/express": "^4.17.1",
+    "body-parser": "^1.19.0",
     "express": "^4.17.1",
     "tslib": "~1.10.0"
   }
