+++ tsconfig.json
@@ -1,9 +1,14 @@
 {
   "compilerOptions": {
-    "target": "es5",
     "module": "commonjs",
-    "strict": true,
+    "esModuleInterop": true,
+    "resolveJsonModule": true,
+    "target": "es5",
+    "noImplicitAny": true,
+    "moduleResolution": "node",
+    "sourceMap": true,
     "outDir": "dist",
-    "sourceMap": true
+    "baseUrl": ".",
+    "removeComments": true
   }
 }
